+++ packages/doc-gen/src/remark-install.ts
@@ -2,7 +2,7 @@ import type { Code, Root } from 'mdast';
 import type { Transformer } from 'unified';
 import { visit } from 'unist-util-visit';
 import convert from 'npm-to-yarn';
-import { createElement } from './utils';
+import { createElement, expressionToAttribute } from './utils';
 
 interface PackageManager {
   name: string;
@@ -16,6 +16,18 @@ interface PackageManager {
 export type RemarkInstallOptions = Partial<{
   Tabs: string;
   Tab: string;
+
+  /**
+   * Persist Tab value (Fumadocs UI only)
+   *
+   * @defaultValue false
+   */
+  persist?:
+    | {
+        id: string;
+      }
+    | false;
+
   packageManagers: PackageManager[];
 }>;
 
@@ -33,6 +45,7 @@ export type RemarkInstallOptions = Partial<{
 export function remarkInstall({
   Tab = 'Tab',
   Tabs = 'Tabs',
+  persist = false,
   packageManagers = [
     { command: (cmd) => convert(cmd, 'npm'), name: 'npm' },
     { command: (cmd) => convert(cmd, 'pnpm'), name: 'pnpm' },
@@ -50,19 +63,35 @@ export function remarkInstall({
 
       const insert = createElement(
         Tabs,
-        {
-          items: {
+        [
+          ...(typeof persist === 'object'
+            ? [
+                {
+                  type: 'mdxJsxAttribute',
+                  name: 'id',
+                  value: persist.id,
+                },
+                {
+                  type: 'mdxJsxAttribute',
+                  name: 'persist',
+                  value: null,
+                },
+              ]
+            : []),
+          expressionToAttribute('items', {
             type: 'ArrayExpression',
             elements: packageManagers.map(({ name }) => ({
               type: 'Literal',
               value: name,
             })),
-          },
-        },
+          }),
+        ],
         packageManagers.map(({ command, name }) => ({
           type: 'mdxJsxFlowElement',
           name: Tab,
-          attributes: [{ type: 'mdxJsxAttribute', name: 'value', value: name }],
+          attributes: [
+            { type: 'mdxJsxAttribute', name: 'value', value: name },
+          ].filter(Boolean),
           children: [
             {
               type: 'code',