+++ packages/mdx-remote/src/github/source.ts
@@ -20,6 +20,13 @@ import { USE_LOCAL } from '@/github/constants';
 export async function createSourceAuto<Frontmatter extends PageData>(
   options: GetFilesOptions & {
     /**
+     * Choose the content source
+     *
+     * @defaultValue 'auto'
+     */
+    from?: 'auto' | 'local' | 'github';
+
+    /**
      * Options when GitHub content source is used
      */
     github: Omit<GetGitHubFilesOptions, keyof GetFilesOptions> &
@@ -33,12 +40,15 @@ export async function createSourceAuto<Frontmatter extends PageData>(
     metaData: MetaData;
   }>
 > {
+  const { from = 'auto', ...rest } = options;
+
   return {
-    files: USE_LOCAL
-      ? await getLocalFiles(options)
-      : await getGitHubFiles({
-          ...options,
-          ...options.github,
-        }),
+    files:
+      (from === 'auto' && USE_LOCAL) || from === 'local'
+        ? await getLocalFiles(rest)
+        : await getGitHubFiles({
+            ...rest,
+            ...rest.github,
+          }),
   };
 }