+++ packages/content-collections/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @fumadocs/content-collections
 
+## 1.0.3
+
+### Patch Changes
+
+- 6e9d4bf: Add `icon` and `root` fields to meta schema
+- 2db6172: Add `cwd` option by default
+
+## 1.0.2
+
+### Patch Changes
+
+- dd1e4f3: Change type definitions of built-in schemas
+  - fumadocs-core@12.4.1
+
+## 1.0.1
+
+### Patch Changes
+
+- 8efc775: Fix Content Collections compatibility
+  - fumadocs-core@12.4.0
+
 ## 1.0.0
 
 ### Major Changes