+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,33 @@
 # next-docs-ui
 
+## 12.4.2
+
+### Patch Changes
+
+- 503e8e9: Improve Object Collaspible
+- Updated dependencies [503e8e9]
+  - fumadocs-core@12.4.2
+
+## 12.4.1
+
+### Patch Changes
+
+- fumadocs-core@12.4.1
+
+## 12.4.0
+
+### Minor Changes
+
+- eb36761: Replace link item `secondary` type with `icon` (backward compatible)
+- eb36761: Support `secondary` property in link items
+- eb36761: Support `button` type link item
+- eb36761: Support `on` filter in link items
+
+### Patch Changes
+
+- 33ffa99: Improve design details
+  - fumadocs-core@12.4.0
+
 ## 12.3.6
 
 ### Patch Changes